run pure bf16, Kahan

+++ anyprecision_optimizer.py
@@ -25,7 +25,7 @@ class AnyPrecisionAdamW(Optimizer):
         eps=1e-8,
         weight_decay=0.0,
         use_kahan_summation=False,
-        momentum_dtype=torch.float32,
+        momentum_dtype=torch.bfloat16,
         variance_dtype=torch.bfloat16,
         compensation_buffer_dtype=torch.bfloat16,
     ):
